+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.2.2] - 2022-09-22
+
 ### Fixed
 
 - Check prefix list entries prior to adding new entry
@@ -117,7 +119,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Helm chart
 
-[Unreleased]: https://github.com/giantswarm/aws-network-topology-operator/compare/v0.2.1...HEAD
+[Unreleased]: https://github.com/giantswarm/aws-network-topology-operator/compare/v0.2.2...HEAD
+[0.2.2]: https://github.com/giantswarm/aws-network-topology-operator/compare/v0.2.1...v0.2.2
 [0.2.1]: https://github.com/giantswarm/aws-network-topology-operator/compare/v0.2.0...v0.2.1
 [0.2.0]: https://github.com/giantswarm/aws-network-topology-operator/compare/v0.1.7...v0.2.0
 [0.1.7]: https://github.com/giantswarm/aws-network-topology-operator/compare/v0.1.6...v0.1.7